updating default fig locations to make room for peths and added some stuff for plotting peths and trials in the time viewer

+++ +jrclust/+views/plotFigTraces.m
@@ -81,7 +81,7 @@ function tracesFilt = plotFigTraces(hFigTraces, hCfg, tracesRaw, resetAxis, hClu
             offset = sum(cellfun(@(hR) hR.nSamples, hRecs));
         end
 
-        recTimes = hClust.spikeTimes - int32(offset);
+        recTimes = hClust.spikeTimes - uint64(offset);
 
         tStart = single(hFigTraces.figData.windowBounds(1) - 1)/hCfg.sampleRate;
         if hCfg.nSegmentsTraces > 1
